refactor(ms-races): refactored the code to remove unused method.
